bug修复

+++ pms/src/main/java/com/upedge/pms/modules/product/entity/ProductVariant.java
@@ -45,7 +45,7 @@ public class ProductVariant{
 	/**
 	 * 变体库存
 	 */
-    private Long inventory;
+    private Integer inventory;
 	/**
 	 * 重量
 	 */
